[border-agent] move border agent ephemeral key APIs to a separate header
This change moves all public APIs related to the Border Agent's
ephemeral key feature out of `border_agent.h` and into a new,
dedicated header file `border_agent_ephemeral_key.h`.

The corresponding C API implementations are also moved from
`border_agent_api.cpp` into a new dedicated
`border_agent_ephemeral_key_api.cpp` file.

This improves the organization and modularity of the public API,
making the codebase easier to navigate and understand.
