+++ include/openthread/BUILD.gn
@@ -43,6 +43,7 @@ source_set("openthread") {
     "backbone_router_ftd.h",
     "ble_secure.h",
     "border_agent.h",
+    "border_agent_ephemeral_key.h",
     "border_agent_tracker.h",
     "border_agent_txt_data.h",
     "border_router.h",
