+++ include/openthread/instance.h
@@ -52,7 +52,7 @@ extern "C" {
  *
  * @note This number versions both OpenThread platform and user APIs.
  */
-#define OPENTHREAD_API_VERSION (561)
+#define OPENTHREAD_API_VERSION (562)
 
 /**
  * @addtogroup api-instance
