+++ src/core/api/border_agent_api.cpp
@@ -126,66 +126,4 @@ otError otBorderAgentEvictActiveCommissioner(otInstance *aInstance)
 }
 #endif
 
-#if OPENTHREAD_CONFIG_BORDER_AGENT_EPHEMERAL_KEY_ENABLE
-
-otBorderAgentEphemeralKeyState otBorderAgentEphemeralKeyGetState(otInstance *aInstance)
-{
-    return MapEnum(AsCoreType(aInstance).Get<MeshCoP::BorderAgent::EphemeralKeyManager>().GetState());
-}
-
-void otBorderAgentEphemeralKeySetEnabled(otInstance *aInstance, bool aEnabled)
-{
-    AsCoreType(aInstance).Get<MeshCoP::BorderAgent::EphemeralKeyManager>().SetEnabled(aEnabled);
-}
-
-otError otBorderAgentEphemeralKeyStart(otInstance *aInstance,
-                                       const char *aKeyString,
-                                       uint32_t    aTimeout,
-                                       uint16_t    aUdpPort)
-{
-    AssertPointerIsNotNull(aKeyString);
-
-    return AsCoreType(aInstance).Get<MeshCoP::BorderAgent::EphemeralKeyManager>().Start(aKeyString, aTimeout, aUdpPort);
-}
-
-void otBorderAgentEphemeralKeyStop(otInstance *aInstance)
-{
-    AsCoreType(aInstance).Get<MeshCoP::BorderAgent::EphemeralKeyManager>().Stop();
-}
-
-uint16_t otBorderAgentEphemeralKeyGetUdpPort(otInstance *aInstance)
-{
-    return AsCoreType(aInstance).Get<MeshCoP::BorderAgent::EphemeralKeyManager>().GetUdpPort();
-}
-
-void otBorderAgentEphemeralKeySetCallback(otInstance                       *aInstance,
-                                          otBorderAgentEphemeralKeyCallback aCallback,
-                                          void                             *aContext)
-{
-    AsCoreType(aInstance).Get<MeshCoP::BorderAgent::EphemeralKeyManager>().SetCallback(aCallback, aContext);
-}
-
-const char *otBorderAgentEphemeralKeyStateToString(otBorderAgentEphemeralKeyState aState)
-{
-    OT_ASSERT(aState <= OT_BORDER_AGENT_STATE_ACCEPTED);
-
-    return MeshCoP::BorderAgent::EphemeralKeyManager::StateToString(MapEnum(aState));
-}
-
-#if OPENTHREAD_CONFIG_VERHOEFF_CHECKSUM_ENABLE
-
-otError otBorderAgentEphemeralKeyGenerateTap(otBorderAgentEphemeralKeyTap *aTap)
-{
-    return AsCoreType(aTap).GenerateRandom();
-}
-
-otError otBorderAgentEphemeralKeyValidateTap(const otBorderAgentEphemeralKeyTap *aTap)
-{
-    return AsCoreType(aTap).Validate();
-}
-
-#endif
-
-#endif // OPENTHREAD_CONFIG_BORDER_AGENT_EPHEMERAL_KEY_ENABLE
-
 #endif // OPENTHREAD_CONFIG_BORDER_AGENT_ENABLE
